requirements.txt
aws-cdk-lib==2.37.1